pyproject.toml
[tool.poetry]
name = "csobclient"
version = "0.0.1"
description = "Library for communication with ČSOB API"
authors = ["Andrii Nechaiev <andrewnech@gmail.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/litteratum/csobclient"
classifiers=[
    "Development Status :: 1 - Planning",
    "Intended Audience :: Developers",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: Implementation :: CPython",
    "Topic :: Software Development :: Libraries",
    "License :: OSI Approved :: MIT License",
    "Operating System :: POSIX :: Linux",
    "Topic :: Utilities",
]

[tool.poetry.dependencies]
python = ">=3.7, <3.11"
requests = "^2.28.2"
pycryptodome = "^3.17"

[tool.poetry.dev-dependencies]
black = "^22.1.0"
pylint = "^2.12.2"
pydocstyle = "^6.1.1"
pytest = "^7.0.1"
pytest-cov = "^3.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
